+++ src/com/intellij/idea/plugin/hybris/settings/CCv2SettingsConfigurableProvider.kt
@@ -18,27 +18,21 @@
 package com.intellij.idea.plugin.hybris.settings
 
 import com.intellij.idea.plugin.hybris.common.equalsIgnoreOrder
+import com.intellij.idea.plugin.hybris.tools.ccv2.strategies.CCv2Strategy
 import com.intellij.idea.plugin.hybris.ui.CCv2SubscriptionListPanel
 import com.intellij.openapi.fileChooser.FileChooserDescriptorFactory
 import com.intellij.openapi.options.BoundSearchableConfigurable
 import com.intellij.openapi.options.ConfigurableProvider
-import com.intellij.openapi.util.SystemInfo
 import com.intellij.ui.components.JBPasswordField
 import com.intellij.ui.components.textFieldWithBrowseButton
-import com.intellij.ui.dsl.builder.AlignX
-import com.intellij.ui.dsl.builder.RowLayout
-import com.intellij.ui.dsl.builder.bindText
-import com.intellij.ui.dsl.builder.panel
-import java.nio.file.Paths
-import kotlin.io.path.exists
-import kotlin.io.path.isExecutable
+import com.intellij.ui.dsl.builder.*
 
-class HybrisCLISettingsConfigurableProvider : ConfigurableProvider() {
+class CCv2SettingsConfigurableProvider : ConfigurableProvider() {
 
     override fun createConfigurable() = SettingsConfigurable()
 
     class SettingsConfigurable : BoundSearchableConfigurable(
-        "CCv2 CLI", "[y] SAP Commerce Cloud CLI configuration."
+        "CCv2", "[y] SAP Commerce Cloud CCv2 configuration."
     ) {
 
         private val appSettings = HybrisApplicationSettingsComponent.getInstance()
@@ -50,11 +44,12 @@ class HybrisCLISettingsConfigurableProvider : ConfigurableProvider() {
         private lateinit var sapCLITokenTextField: JBPasswordField
         private val ccv2SubscriptionListPanel = CCv2SubscriptionListPanel(originalCCv2Subscriptions)
 
+        // TODO: we may use SAP CCM CLI or native REST API in the future
         override fun createPanel() = panel {
             row {}.comment(
                 """
-                All details on using SAP CCM can be found in official documentation <a href="https://help.sap.com/docs/SAP_COMMERCE_CLOUD_PUBLIC_CLOUD/9116f1cfd16049c3a531bfb6a681ff77/8acde53272c64efb908b9f0745498015.html">help.sap.com - Command Line Interface</a>.
-            """.trimIndent()
+                    All details on using SAP CCM can be found in official documentation <a href="https://help.sap.com/docs/SAP_COMMERCE_CLOUD_PUBLIC_CLOUD/9116f1cfd16049c3a531bfb6a681ff77/8acde53272c64efb908b9f0745498015.html">help.sap.com - Command Line Interface</a>.
+                """.trimIndent()
             )
 
             row {
@@ -68,27 +63,17 @@ class HybrisCLISettingsConfigurableProvider : ConfigurableProvider() {
                 )
                     .comment(
                         """
-                        SAP Commerce Cloud command line interface installation directory.<br>
-                        Choose directory extracted from the <strong>CXCOMMCLI00P_*.zip</strong> file to enable CCv2 CLI integration. 
-                    """.trimIndent()
+                            SAP Commerce Cloud command line interface installation directory.<br>
+                            Choose directory extracted from the <strong>CXCOMMCLI00P_*.zip</strong> file to enable CCv2 CLI integration. 
+                        """.trimIndent()
                     )
                     .bindText(
                         { state.sapCLIDirectory ?: "" },
                         { state.sapCLIDirectory = it }
                     )
                     .validationOnInput {
-                        val executable = if (SystemInfo.isWindows) "sapccm.bat"
-                        else "sapccm"
-
-                        val valid = Paths.get(it.text, "bin", executable)
-                            .takeIf { path -> path.exists() }
-                            ?.isExecutable()
-                            ?: false
-                        if (!valid) {
-                            this.error("Invalid SAP CCM directory, cannot find <strong>bin/$executable</strong> executable file.")
-                        } else {
-                            null
-                        }
+                        CCv2Strategy.getSAPCCMCCv2Strategy().validateSAPCCMDirectory(it.text)
+                            ?.let { message -> this.error(message) }
                     }
                     .align(AlignX.FILL)
             }.layout(RowLayout.PARENT_GRID)
@@ -98,9 +83,9 @@ class HybrisCLISettingsConfigurableProvider : ConfigurableProvider() {
                 sapCLITokenTextField = passwordField()
                     .comment(
                         """
-                        Specify developer specific Token for CCv2 CLI, it will be stored in the OS specific secure storage under <strong>SAP CX CLI Token</strong> alias.<br>
-                        Official documentation <a href="https://help.sap.com/docs/SAP_COMMERCE_CLOUD_PUBLIC_CLOUD/0fa6bcf4736c46f78c248512391eb467/b5d4d851cbd54469906a089bb8dd58d8.html">help.sap.com - Generating API Tokens</a>.
-                    """.trimIndent()
+                            Specify developer specific Token for CCv2 CLI, it will be stored in the OS specific secure storage under <strong>SAP CX CLI Token</strong> alias.<br>
+                            Official documentation <a href="https://help.sap.com/docs/SAP_COMMERCE_CLOUD_PUBLIC_CLOUD/0fa6bcf4736c46f78c248512391eb467/b5d4d851cbd54469906a089bb8dd58d8.html">help.sap.com - Generating API Tokens</a>.
+                        """.trimIndent()
                     )
                     .align(AlignX.RIGHT)
                     .onIsModified { originalSAPCLIToken != String(sapCLITokenTextField.password) }
@@ -108,7 +93,7 @@ class HybrisCLISettingsConfigurableProvider : ConfigurableProvider() {
                         sapCLITokenTextField.isEnabled = false
 
                         appSettings.loadSAPCLIToken {
-                            val sapCLIToken = appSettings.sapCLIToken
+                            val sapCLIToken = appSettings.ccv2Token
                             originalSAPCLIToken = sapCLIToken
 
                             sapCLITokenTextField.text = sapCLIToken
@@ -123,6 +108,17 @@ class HybrisCLISettingsConfigurableProvider : ConfigurableProvider() {
                     .align(AlignX.FILL)
                     .component
             }.layout(RowLayout.PARENT_GRID)
+
+            row {
+                label("Connection timeout:")
+                intTextField(10..Int.MAX_VALUE)
+                    .comment(
+                        """
+                            Indicates waiting time in seconds when invoking SAP CCM.
+                        """.trimIndent()
+                    )
+                    .bindIntText(state::sapCLITimeout)
+            }
 
             group("CCv2 Subscriptions", false) {
                 row {