+++ src/com/intellij/idea/plugin/hybris/tools/ccv2/strategies/SAPCCMCCv2Strategy.kt
@@ -16,40 +16,43 @@
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  */
 
-package com.intellij.idea.plugin.hybris.tools.cli
+package com.intellij.idea.plugin.hybris.tools.ccv2.strategies
 
 import com.intellij.execution.ExecutionException
 import com.intellij.execution.configurations.GeneralCommandLine
 import com.intellij.execution.configurations.JavaCommandLineStateUtil
 import com.intellij.execution.process.CapturingProcessAdapter
 import com.intellij.execution.process.ProcessEvent
+import com.intellij.ide.BrowserUtil
+import com.intellij.idea.plugin.hybris.common.HybrisConstants
+import com.intellij.idea.plugin.hybris.notifications.Notifications
+import com.intellij.idea.plugin.hybris.settings.CCv2SettingsConfigurableProvider
 import com.intellij.idea.plugin.hybris.settings.CCv2Subscription
 import com.intellij.idea.plugin.hybris.settings.HybrisApplicationSettingsComponent
-import com.intellij.idea.plugin.hybris.tools.cli.dto.CCv2Environment
-import com.intellij.idea.plugin.hybris.tools.cli.dto.CCv2EnvironmentType
+import com.intellij.idea.plugin.hybris.tools.ccv2.dto.CCv2Environment
+import com.intellij.idea.plugin.hybris.tools.ccv2.dto.CCv2EnvironmentType
+import com.intellij.notification.NotificationType
 import com.intellij.openapi.components.Service
+import com.intellij.openapi.options.ShowSettingsUtil
 import com.intellij.openapi.progress.ProgressManager
 import com.intellij.openapi.project.Project
 import com.intellij.openapi.util.SystemInfo
-import com.intellij.util.messages.Topic
 import java.io.File
 import java.nio.charset.StandardCharsets
+import java.nio.file.Paths
+import kotlin.io.path.exists
+import kotlin.io.path.isExecutable
 
-@Service(Service.Level.PROJECT)
-class CCv2CLIService(val project: Project) {
+@Service
+class SAPCCMCCv2Strategy : CCv2Strategy {
 
-    private val messageBus = project.messageBus
-
-    fun fetchEnvironments(subscriptions: Collection<CCv2Subscription>): Map<CCv2Subscription, Collection<CCv2Environment>> {
-        messageBus.syncPublisher(TOPIC_ENVIRONMENT).fetchingStarted()
-
-        val indicator = ProgressManager.getInstance().progressIndicator
+    override fun fetchEnvironments(project: Project, ccv2Token: String, subscriptions: Collection<CCv2Subscription>): Map<CCv2Subscription, Collection<CCv2Environment>> {
         val appSettings = HybrisApplicationSettingsComponent.getInstance()
-        val cliToken = appSettings.sapCLIToken ?: return emptyMap()
-        val authParameters = arrayOf("config", "set", "auth-credentials", cliToken)
+        val indicator = ProgressManager.getInstance().progressIndicator
+        val authParameters = arrayOf("config", "set", "auth-credentials", ccv2Token)
 
         indicator.text2 = "Authenticating with the provided token..."
-        execCLI(appSettings, *authParameters) ?: return emptyMap()
+        executeCLI(project, appSettings, *authParameters) ?: return emptyMap()
 
         return subscriptions
             .associateWith { subscription ->
@@ -58,7 +61,7 @@ class CCv2CLIService(val project: Project) {
                 // the first row is a header, second+ rows are data rows
                 // we need to identify start position for each column by corresponding Column header
                 indicator.text2 = "Fetching environments for subscription: $subscription"
-                val processOutput = execCLI(appSettings, *parameters)
+                val processOutput = executeCLI(project, appSettings, *parameters)
                     ?.takeIf { it.size > 1 }
                     ?: return@associateWith emptyList()
 
@@ -80,10 +83,22 @@ class CCv2CLIService(val project: Project) {
                         )
                     }
             }
-            .also { messageBus.syncPublisher(TOPIC_ENVIRONMENT).fetchingCompleted(it) }
     }
 
-    private fun execCLI(
+    fun validateSAPCCMDirectory(directory: String): String? {
+        val executable = if (SystemInfo.isWindows) "sapccm.bat"
+        else "sapccm"
+
+        val valid = Paths.get(directory, "bin", executable)
+            .takeIf { path -> path.exists() }
+            ?.isExecutable()
+            ?: false
+        return if (!valid) "Invalid SAP CCM directory, cannot find <strong>bin/$executable</strong> executable file."
+        else null
+    }
+
+    private fun executeCLI(
+        project: Project,
         appSettings: HybrisApplicationSettingsComponent,
         vararg parameters: String
     ): List<String>? {
@@ -106,34 +121,59 @@ class CCv2CLIService(val project: Project) {
             val handler = JavaCommandLineStateUtil.startProcess(gcl, true)
             handler.addProcessListener(object : CapturingProcessAdapter() {
                 override fun processTerminated(event: ProcessEvent) {
-                    if (event.exitCode != 0) {
-                        // TODO: baaad, do something...
-                    } else {
+                    if (event.exitCode == 0) {
                         processOutput.addAll(output.stdoutLines)
+                    } else {
+                        processOutput.add(output.stderr)
+                        processOutput.add(output.stdout)
                     }
                 }
             })
 
             handler.startNotify()
 
-            // TODO: add timeout configuration
-            val waitFor = handler.waitFor(10 * 1000L)
+            val timeout = appSettings.state.sapCLITimeout * 1000L
+            val waitFor = handler.waitFor(timeout)
             if (!waitFor) {
-                // TODO: inform somehow
                 handler.destroyProcess()
+
+                Notifications
+                    .create(
+                        NotificationType.WARNING,
+                        "SAP CCM: Interrupted on timeout",
+                        "Exceeded current timeout of $timeout seconds, it can be adjusted within CCv2 settings."
+                    )
+                    .addAction("Open Settings") { _, _ ->
+                        ShowSettingsUtil.getInstance().showSettingsDialog(project, CCv2SettingsConfigurableProvider.SettingsConfigurable::class.java)
+                    }
+                    .hideAfter(10)
+                    .notify(project)
+
                 return null
             }
 
-            return if (handler.exitCode != 0) null
-            else processOutput
+            return if (handler.exitCode != 0) {
+                val content = processOutput.joinToString(System.lineSeparator())
+                val notification = Notifications
+                    .create(
+                        NotificationType.WARNING,
+                        "SAP CCM: Unable to process request",
+                        content
+                    )
+                if (content.contains("UNAUTHORIZED")) {
+                    notification
+                        .addAction("Open Settings") { _, _ ->
+                            ShowSettingsUtil.getInstance().showSettingsDialog(project, CCv2SettingsConfigurableProvider.SettingsConfigurable::class.java)
+                        }
+                        .addAction("Generating API Tokens...") { _, _ -> BrowserUtil.browse(HybrisConstants.URL_HELP_GENERATING_API_TOKENS) }
+                }
+                notification
+                    .notify(project)
+
+                return null
+            } else processOutput
         } catch (e: ExecutionException) {
             return null
         }
-    }
-
-    companion object {
-
-        val TOPIC_ENVIRONMENT = Topic("HYBRIS_CCv2_CLI_ENVIRONMENT_LISTENER", CCv2CLIEnvironmentListener::class.java)
-        fun getInstance(project: Project): CCv2CLIService = project.getService(CCv2CLIService::class.java)
     }
 }